Make migrations idempotent Serves me right for doing other work...

+++ database/migrations/2024_12_23_131406_drop_division_from_rank_actions_table.php
@@ -11,9 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('rank_actions', function (Blueprint $table) {
-            $table->dropColumn('division_id');
-        });
+        if (Schema::hasColumn('rank_actions', 'division_id')) {
+            Schema::table('rank_actions', function (Blueprint $table) {
+                $table->dropColumn('division_id');
+            });
+        }
+
     }
 
     /**
