+++ database/migrations/2024_12_23_131648_drop_expires_at_column_from_award_member_table.php
@@ -4,16 +4,17 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('award_member', function (Blueprint $table) {
-            $table->dropColumn('expires_at');
-        });
+        if (Schema::hasColumn('award_member', 'expires_at')) {
+            Schema::table('award_member', function (Blueprint $table) {
+                $table->dropColumn('expires_at');
+            });
+        }
     }
 
     /**
